refactor(transform): split into check/status + install_hook
commands/transform.rs (957 LOC code / 253 LOC tests, 1559 total)
bundled three sub-features that share a CLI grouping but no logic:

- `check` + `status` — review the deployed/source/baseline triangle
   and reverse-merge unambiguous edits back to source. Tightly
   coupled (status is the read-only view of check's classification).
- `install_hook` + `hook_is_installed` + `managed_block` +
   `find_managed_block` — git pre-commit hook file management
   (HOOK_GUARD_* constants, InstallHookOutcome enum,
   InstallHookResult struct).

These two halves share zero types, helpers, or call paths. Tests
already clustered by section (`// ── status`, `// ── install_hook`,
`// ── hook upgrade`).

Now:

- transform/mod.rs           — types + status + check + helpers,
                                check/status tests inline
- transform/install_hook.rs  — hook installer + its types + its tests
- transform/test_support.rs  — shared `make_ctx` (pub(super))

Public API preserved via re-exports: `commands::transform::check`,
`status`, `install_hook`, `hook_is_installed`, `managed_block`,
`InstallHookOutcome`, `InstallHookResult`, `TransformCheckResult`,
`TransformStatusResult` all still resolve.

Same 1140 tests pass. mod.rs goes from 1559 → 1049, with the
install-hook surface (~500 LOC) now in its own focused file.
